Split up QueryData classes into separate files.
