+++ mikeio1d/result_query/__init__.py
@@ -1,0 +1,6 @@
+from .query_data import QueryData
+from .query_data_catchment import QueryDataCatchment
+from .query_data_global import QueryDataGlobal
+from .query_data_node import QueryDataNode
+from .query_data_reach import QueryDataReach
+from .query_data_structure import QueryDataStructure
